requirements.txt
numpy>=1.23.4
pandas==1.5.3
databricks-connect==13.3.0
matplotlib==3.4.2
pmdarima==2.0.3
pystan
prophet
hyperopt==0.2.7
chardet